Figured out some migration things
So, in the reference object, there's a field that is basically a
boolean of whether or not the object is being migrated. Since
everything will be private, they have to use the built in methods on
the reference object to get information. It gives us a lot of control,
actually.